requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
fastapi==0.121.1
graphql-core==3.2.7
h11==0.16.0
idna==3.11
lia-web==0.2.3
packaging==25.0
psycopg2-binary==2.9.11
pydantic==2.12.4
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlmodel==0.0.27
starlette==0.49.3
strawberry-graphql==0.285.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
